Split command and options
