+++ cli/src/index.ts
@@ -9,9 +9,17 @@
  ********************************************************************************/
 
 export * from './create-namespace';
+export * from './create-namespace-options';
 export * from './get';
+export * from './get-options';
+export * from './login';
+export * from './login-options';
+export * from './logout';
 export * from './publish';
+export * from './publish-options';
 export * from './registry';
+export * from './registry-options';
 export * from './verify-pat';
+export * from './verify-pat-options';
 export { isLicenseOk } from './check-license';
 export { validateManifest, readManifest } from './util';
